Move things around for a neater project structure
