+++ app/components/welcome_banner/welcome_banner.stories.js
