+++ app/isomorphic_helpers.js
